requirements.txt
accelerate==1.5.2
aiohttp-cors==0.7.0
beautifulsoup4==4.13.3
colorful==0.5.6
evaluate==0.4.3
flax==0.10.3
google-cloud-aiplatform==1.82.0
groq==0.18.0
grpcio-status==1.70.0
httptools==0.6.4
importlib-resources==6.5.2
ipykernel==6.29.5
matplotlib==3.10.1
nltk==3.9.1
opencensus==0.11.4
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
pip-chill==1.0.3
polars==1.23.0
py-spy==0.4.0
python-dotenv==1.0.1
replicate==1.0.4
scikit-learn==1.6.1
smart-open==7.1.0
tf-keras==2.18.0
uvloop==0.21.0
virtualenv==20.29.1
vllm==0.6.6.post1
watchfiles==1.0.4
websockets==14.2
